refactor: Remove nMyStartingHeight from CNode/Connman
Summary:
```
CNode and CConnman keep track of the active chain height when CNodes have been created, but apart from serializing the int once (when sending a version message), it is unused. So it can simply be removed in favor of a single int in PeerMan that can do the same.
```

Backport of [[https://github.com/bitcoin/bitcoin/pull/20649 | core#20649]].

Depends on D10884.

Ref T1696.

Test Plan:
  ninja all check-all
  ninja bitcoin-fuzzers

Reviewers: #bitcoin_abc, PiRK

Reviewed By: #bitcoin_abc, PiRK

Subscribers: PiRK

Maniphest Tasks: T1696

Differential Revision: https://reviews.bitcoinabc.org/D10885