[CashTab] Fix TerserPlugin and other npm packages with vulnerabilities
Summary: This diff fix TerserPlugin package with vulnerability and apply necessary changes due to the new version. Also applies changes to a number of packages to remove vulnerabilities.

Test Plan:
On master, run "npm audit", you should see 7 vulnerabilities identified. Upon applying this patch, the total should come down to 2: one from bch-js which will be addressed in their repo since it is a breaking change. The other from babel-plugin-fbt-runtime which needs further investigation, but it is rated as a low vulnerability.

Run "npm install" and "npm start" to check if the app is working properly.

Reviewers: #bitcoin_abc, josephroyking

Reviewed By: #bitcoin_abc, josephroyking

Subscribers: josephroyking

Differential Revision: https://reviews.bitcoinabc.org/D8808