[CASHTAB] Remove the .prettierignore file
Summary:
This file is used to exclude files from being overriden when prettier is
used with the `--write` option. It is not used to exclude files from
being parsed by the linter.
Since we use arcanist autofix feature to reformat the files, there is no
need for this file.

Test Plan:
  arc lint --everything

Reviewers: #bitcoin_abc, josephroyking

Reviewed By: #bitcoin_abc, josephroyking

Subscribers: josephroyking

Differential Revision: https://reviews.bitcoinabc.org/D8823