[CASHTAB] Update the jest dependency
Summary:
Update `jest` to the latest version. This will allow for creating
coverage report files as needed, which is not supported with current
version.
Note that the `collectCoverageFrom` regex exclusion has been removed. It
was not matching any file and is very likely broken:
https://github.com/facebook/jest/issues/9464, so it's better to no rely
on it.
The coverage behavior seems to change quite a lot, because it takes into
acount the files with 0 coverage which were previously not accounted
for.

Test Plan:
  npm install
  npm test
  npm run test:coverage

Reviewers: #bitcoin_abc, deadalnix, majcosta

Reviewed By: #bitcoin_abc, deadalnix, majcosta

Subscribers: josephroyking, deadalnix

Differential Revision: https://reviews.bitcoinabc.org/D8828