[CI] Improve cashtab reporting in case of a test failure
Summary:
This diff adds a default (text) reporter to make the tests log available
in the build log, and traps the junit report move to the build directory
so that it gets available even if an error stops the build.

Test Plan:
Cause a test to fail, then:
  ./contrib/teamcity/build-configurations.py cashtab-tests
Check the failure is logged to the console and to the junit report.

Reviewers: #bitcoin_abc, majcosta

Reviewed By: #bitcoin_abc, majcosta

Subscribers: majcosta

Differential Revision: https://reviews.bitcoinabc.org/D8842