[CI] Work around LCOV stupid behavior
Summary:
D8874 was tested using a recent lcov version. Previous version are
stupid enough to print their output to stderr (when no error) instead of
stdout.
This diff works around this by redirecting stderr to stdout, so the
coverage-summary.txt file is not empty.

Test Plan:
On Debian:
  ./contrib/teamcity/build-configurations.py cashtab-tests
Check the coverage summary has content in it.

Reviewers: #bitcoin_abc, deadalnix

Reviewed By: #bitcoin_abc, deadalnix

Differential Revision: https://reviews.bitcoinabc.org/D8877