GUI: Write PSBTs to file with binary mode
Summary:
PSBT files should be opened in binary mode as on windows, all newlines are turned into CRLF which produces invalid PSBTs.

This is a partial backport of [[https://github.com/bitcoin-core/gui/pull/188 | core-gui#188]]

The change to psbtoperationsdialog.cpp is not applicable due to missing backports (core#18027)

Test Plan: Made sure that I can still save and load PSBT file in bitcoin-qt on Linux. I don't have a windows computer to compile and test this, but the change is trivial enough that it should not be necessary.

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Differential Revision: https://reviews.bitcoinabc.org/D11984