[Cashtab] Default wallet name to 1899 address
Summary: When the user creates a new wallet in Cashtab, the default name is the first 5 characters of the wallet address. This is still set to the legacy path (145). This diff sets the default name to the first 5 chars of the Path1899 address, which is the displayed address.

Test Plan: `npm start`, navigate to settings, click 'New Wallet', observe the name of the new wallet, Activate the new wallet, observe that the displayed address on the wallet page matches the wallet's name.

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Differential Revision: https://reviews.bitcoinabc.org/D8983