[CI] Elect more builds to run on diff depending on simple heuristic
Summary:
This diff makes the `build-bench` to run when a diff changes a file in
`src/bench/` and the `secp256k1` builds when a diff changes a file in
`src/secp256k1/`.

Test Plan:
This can't be tested before landed, but you can check that the similar
D8943 is working as expected.

Reviewers: #bitcoin_abc, majcosta

Reviewed By: #bitcoin_abc, majcosta

Differential Revision: https://reviews.bitcoinabc.org/D8997