[Cashtab] Reject transactions lower than backend dust limit
Summary: Cashtab minimum send is currently set to 0.00001 BCHA. The theoretical minimum of the backend is 546 satoshis, or 0.00000546. This diff rejects attempted sends before the tx is broadcast to the backend if they are for less than the backend dust limit, and adds a unit test. Because the app dust limit is still higher than 0.00000546, there is no impact on the app functionality. This is a preliminary step for matching the app limit to the backend limit.

Test Plan: `npm test`

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Differential Revision: https://reviews.bitcoinabc.org/D9018