zmq test: accept arbitrary sequence start number in ZMQSubscriber
Summary:
The ZMQSubscriber reception methods currently assert that the first
received publisher message has a sequence number of zero. In order to
fix the current test flakiness via "syncing up" to nodes in the setup
phase, we have to cope with the situation that messages get lost and the
first actual received message has a sequence number larger than zero.

This is a backport of [[https://github.com/bitcoin/bitcoin/pull/21008 | core#21008]] [2/3]
https://github.com/bitcoin/bitcoin/pull/21008/commits/8666033630eeaf851ec69e018bb53eb23093f4b9

Depends on D10589

Test Plan: `ninja check-functional`

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Differential Revision: https://reviews.bitcoinabc.org/D10590