[Cashtab] replace BCHA decimal unit with a constant
Summary: BCHA has 8 decimal places, and the value "8" is occasionally used in calculation that convert between BCHA and satoshis. This diff replaces "8" with a constant in `Ticker.js` to support potential adjustments to this value. Note: this diff does not completely prepare the app to support different decimal places. An additional diff is required that would remove use of the `.toSatoshis` function from `bch-js` and replace it with a generalized conversion function.

Test Plan: `npm test`

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Subscribers: Fabien

Differential Revision: https://reviews.bitcoinabc.org/D9054