[Cashtab patch] Adjust input field step for cash or crypto
Summary: The `step` parameter of an input field generates HTML errors if user enters a value that is not a multiple of the step. Set this value to the minimum base unit of crypto or cash.

Test Plan: `npm test`. This change only impacts an HTML error that can appear when the user successfully sends a transaction at an amount that is not a multiple of the `step` parameter.

Reviewers: #bitcoin_abc, PiRK

Reviewed By: #bitcoin_abc, PiRK

Differential Revision: https://reviews.bitcoinabc.org/D9089