[net processing] Move RelayTransaction into PeerManager
Summary:
```
It moves the RelayTransaction() function to be a member function of the PeerManager class. This is required in order to move the transaction inventory data into the Peer object, since Peer objects are only accessible from within PeerManager.
```

Backport of [[https://github.com/bitcoin/bitcoin/pull/21162 | core#21162]].

Test Plan:
  ninja all check-all

Reviewers: #bitcoin_abc, PiRK

Reviewed By: #bitcoin_abc, PiRK

Differential Revision: https://reviews.bitcoinabc.org/D12204