[Cashtab] replace legacy DUST param with constant
Summary:
Cashtab has an artefact parameter `DUST` leftover from mint.bitcoin.com days. This diff removes the parameter and replaces it with the constant `currency.dust`. A unit test is added to ensure this new variable is parsed the same way as the previous one.

This improvement is a step toward refactoring all functions that deal with the smallest base unit to use the `BigNumber.js` library. This will prepare the app to accept arbitrary base units.

Test Plan: `npm test`

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Subscribers: Fabien

Differential Revision: https://reviews.bitcoinabc.org/D9102