style: rename hSocket to sock
Summary:
In the arguments of `InterruptibleRecv()`, `Socks5()` and
`ConnectThroughProxy()` the variable `hSocket` was previously of type
`SOCKET`, but has been changed to `Sock`. Thus rename it to `sock` to
imply its type, to distinguish from other `SOCKET` variables and to
abide to the coding style wrt variables' names.

This is a backport of [[https://github.com/bitcoin/bitcoin/pull/20788 | core#20788]] [5/5]
https://github.com/bitcoin/bitcoin/pull/20788/commits/7bd21ce1efc363b3e8ea1d51dd1410ccd66820cb

Depends on D11006

Test Plan: `ninja all check-all`

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Subscribers: Fabien

Differential Revision: https://reviews.bitcoinabc.org/D11007