[CI] Run the diff tests if any CMakeLists.txt file is changed
Summary:
The cmake list files are likely to impact the build, so make it trigger
the tests. Noticed it was missing from D9217.

Test Plan:
Cannot be tested before it lands, so proof reading only.
Run any CI build using the `diff-node` template to check the yaml file
is not broken.

Reviewers: #bitcoin_abc, PiRK

Reviewed By: #bitcoin_abc, PiRK

Differential Revision: https://reviews.bitcoinabc.org/D9218