[CI] Integrate nanobench with Teamcity
Summary:
The new nanobench benchmarking framework introduced in Dxxxx has no
built-in output format compatible with Teamcity. But since it provides a
very complete json output, it is easy to parse it and generate the
messages we want for teamcity.

This diff makes the `bench-bitcoin` target generate csv and json output
files, and adds a simple python script to parse the json output and send
teamcity messages when run on CI.

Depends on D9226.

Test Plan:
  ninja bench-bitcoin

  ../contrib/teamcity/build-configurations.py build-bench
Check the csv and json files are in the artifacts, and it echoes a lot
of teamcity messages with the correct values.

Reviewers: #bitcoin_abc, PiRK

Reviewed By: #bitcoin_abc, PiRK

Differential Revision: https://reviews.bitcoinabc.org/D9227