build: Remove spaces from variable-printing rules
Summary:
This simplifies parsing when using these rules from scripts.

This is a backport of [[https://github.com/bitcoin/bitcoin/pull/21304 | core#21304]]
https://github.com/bitcoin/bitcoin/pull/21304/commits/84912d4b24382ae022da3a863bd6caa2b8948d94

I was wondering why I had to add this space to find the SDK, when I initially tested guix.

Test Plan:
With the Apple SDK in the expected directory:

`HOSTS=x86_64-apple-darwin contrib/guix/guix-build`

Check that it does not fail early because of not finding the SDK.

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Differential Revision: https://reviews.bitcoinabc.org/D13942