[Cashtab] Improve nav buttons
Summary: Reported issue that the nav buttons in bottom footer are glitchy. The responsive click area is not consistent with the `cursor:hover` display. This is because they were implemented as `TabPanes` in `antd.` This diff removes all `Tab` and `TabPane` components + associated css over-rides and replaces with `button` components, which have a responsive click area that's easier to control.

Test Plan: This is a pure `css` change. `npm start` and verify that buttons look okay. Click between them. Confirm functional and no "misclick" issues as with current app (where cursor appears as pointer but does not activate button). `npm test` to confirm no broken unit tests.

Reviewers: #bitcoin_abc, majcosta

Reviewed By: #bitcoin_abc, majcosta

Subscribers: majcosta

Differential Revision: https://reviews.bitcoinabc.org/D9282