Use correct length sig in sigcheck test
Summary:
Even though the test currently passes (due to dummysigchecker), it's not using the sig + sighashtype
as expected. When debugging, it becomes clear that the sighashtype is being stripped off as part of `OP_CHECKSIG`
despite it not being present. While this doesn't produce any unexpected behavior today, it's confusing and not
technically correct so best to fix before it becomes an issue.

Shoutout to Tobias for finding this.

Test Plan: `ninja check`

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Differential Revision: https://reviews.bitcoinabc.org/D9293