[Cashtab] Replace commas as thousands separator with space
Summary:
Large balances should be displayed with some kind of a thousands separator. This is frequently a comma, as in "1,000,000"

However, some countries use the comma as a decimal separator. A simple space, as in "1 000 000" is more universal. Regional customization would be best, but that can wait.

This diffs implements a space as the default thousands separator and adds a unit test to confirm separator is not added to numbers after the decimal place.

Test Plan: `npm test`

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Differential Revision: https://reviews.bitcoinabc.org/D9287