build: set Unicode true for NSIS installer
Summary:
```
Now that we are using Focal for gitian builds, and have NSIS 3.0+
available (also in Guix), we can create installers that support unicode.

Unicode is only becoming the default beginning with the 3.07 release,
https://nsis.sourceforge.io/Docs/AppendixF.html#v3.07-cl,
so we need to set this attribute to get support.

Should close: #13817
```

On Debian 10 the nsis version is 3.04, so we should be able to use the
feature.

Backport of [[ https://github.com/bitcoin/bitcoin/pull/21333/ | core#21333 ]]

Test Plan: Run the windows gitian build twice, ensure it's deterministic.

Reviewers: #bitcoin_abc, PiRK

Reviewed By: #bitcoin_abc, PiRK

Differential Revision: https://reviews.bitcoinabc.org/D9318