Fix setting CXXFLAGS in linux build docs
Summary:
I needed this while compiling on a resource-restricted machine. It turns out
CXXFLAGS is the environment variable name, not the cmake define name.

Test Plan:
Before this patch, following these instructions gets you a warning
that CXXFLAGS is not used in this project. After this patch, it works as expected.

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Differential Revision: https://reviews.bitcoinabc.org/D9301