[avalanche] add an upper limit for the number of UTXOs in a proof
Summary:
We probably need such a limit because the proof will be exchanged over the network when avalanche nodes meet each other.

The test is done when adding an avalanche peer with the `addavalanchenode` RPC command and when an `AVAHELLO` message is received. Note that right now the AVAHELLO message does not contain the proof, but this will need to be implemented in a future diff.

One consideration should be that this number of UTXOs allows miner to use coinbases to reach the minimum stake amount. But this minimum amount is currently not decided.
Another consideration is the maximum message size of 2MB that can be transmitted over the network to other nodes. This permits a maximum of 11 000 UTXOs.

Test Plan: `ninja all check-all`

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Subscribers: Fabien

Differential Revision: https://reviews.bitcoinabc.org/D9306