[Cashtab] Add theming to better organize color changes
Summary: Remove color specification of `.anticon` class from `App.css`. Replace with `styled-components` `<ThemeProvider>` object and theme. For now, this is only one color affecting only the icons impacted by removing this 1 line of `css.` Will be expanded on so that all app styles can be centrally controlled, as with the parameters in `Ticker.js`.

Test Plan: `npm start` and confirm the icons on the Settings page and the Send page look the same as at `cashtabapp.com`

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Differential Revision: https://reviews.bitcoinabc.org/D9308