mempool: Removed the ancestor_score index
Summary:
```
This is no longer being used (it was used previously for mining). After
the CPFP-removal from the mining code in a previous commit, this index
is no longer needed.

Also removed the comparator CompareTxMemPoolEntryByAncestorFee and the tests for this index and comparator.
```

Port of [[https://gitlab.com/bitcoin-cash-node/bitcoin-cash-node/merge_requests/1113 | bchn#1113]].

Depends on D13122.

Test Plan:
  ninja all check

Reviewers: #bitcoin_abc, PiRK, sdulfari

Reviewed By: #bitcoin_abc, PiRK, sdulfari

Differential Revision: https://reviews.bitcoinabc.org/D13123