[avalanche] refactor conditional processing in ProcessMessage
Summary:
In `ProcessMessage`, when checking for avalanche message types, there
are additional checks made in the same `if` statement. This can cause
the `Unknown command...` log message to be printed at the end of
`ProcessMessage` when the message is not unknown.

This has the additional benefit of factoring some repeated checks.

Test Plan: `ninja all check-all`

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Subscribers: Fabien

Differential Revision: https://reviews.bitcoinabc.org/D9338