[Cashtab] Add latest features to browser extension
Summary:
The extension must be manually built and submitted for review at the Google chrome extension store. The extension uses almost all the same files as the web app. However, the css is slightly different to account for the size of the window pane. `App.js` also includes a link to open the extension in tab instead of the BitcoinABC logo in the top right.

This diff modifies the extension files to support webapp features added over the last month, including themeable css and transaction history.

Test Plan:
1. `npm run extension`
2. Navigate to `chrome://extensions` in the Chrome browser, or `brave://extensions` in Brave browser (Firefox is not yet supported)
3. Enable development mode and load `web/cashtab/dis` as an unpacked extension
4. Open the extension and verify that the app behaves correctly (tx history loads, send screen loads, setting screen loads)
5. Verify that sending and receiving transactions works

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Subscribers: Fabien

Differential Revision: https://reviews.bitcoinabc.org/D9359