[avalanche] add an inventory type and a message for proofs
Summary:
This diff carves out an uncontroversial part of D9327: the definition
of a net message and an inventory type that can be used to exchange
avalanche proofs over the network.

A basic proof object and a msg_avahello are also added to the test
framework, with a simple serialization round trip unit test.
The AvalancheProof object will be later expanded (the code already
exists in abandoned diff D9341)

Test Plan: `ninja && ninja check check-functional`

Reviewers: #bitcoin_abc, deadalnix

Reviewed By: #bitcoin_abc, deadalnix

Subscribers: deadalnix

Differential Revision: https://reviews.bitcoinabc.org/D9360